Use TimedGridWidgetDelegate to move code to submodule
